c304: extract RoutedTab + ROUTED_LABEL + ROUTED_DESC + ROUTED_COLOR to own file (#441 2.1)
Workspace.tsx 6435 LOC -> 6162 LOC (-273 LOC). Entry chunk 167.90 KB -> 163.41 KB.
22 tabs extracted + lazy. typecheck clean, 44/44 tests pass.
